Rework marks, move i3 single-spill into script
Spilling a single container into its grandparent no longer requires
marking it, and any marks on the container will no longer be erased.

Reworks mark handling in and out of script to always use `mark --add`,
and add mappings for clearing marks. This fixes the other problem in #2.

Fixes mappings when swapping windows.

Resolves #2